add support for SECRET_KEY / Django assembler / fix add_args

+++ src/py/gopythongo/assemblers/django.py
@@ -1,4 +1,5 @@
 # -* encoding: utf-8 *-
+import base64
 import shutil
 import os
 
@@ -23,21 +24,28 @@ class DjangoAssembler(BaseAssembler):
     def assembler_type(self) -> str:
         return BaseAssembler.TYPE_ISOLATED
 
-    def add_parser(self, parser: configargparse.ArgumentParser) -> None:
+    def add_args(self, parser: configargparse.ArgumentParser) -> None:
         gr_django = parser.add_argument_group("Django Assembler options")
         gr_django.add_argument("--collect-static", dest="collect_static", action="store_true",
-                               help="run 'django-admin.py collectstatic' inside the bundle")
+                               help="If set, run 'django-admin.py collectstatic' inside the bundle")
         gr_django.add_argument("--static-relative-paths", dest="static_relative",
                                default=False, action="store_true",
-                               help="write relative paths to the resulting static content .tar.gz archive")
+                               help="Write relative paths to the resulting static content .tar.gz archive (UNUSED)")
         gr_django.add_argument("--static-root", dest="static_root", default="static/",
-                               help="where to collect static files from (Django's STATIC_ROOT)")
+                               help="Where to collect static files from (Django's STATIC_ROOT)")
         gr_django.add_argument("--assert-static-root-empty", dest="fresh_static", action="store_true",
-                               help="if set, this script will make sure that STATIC_ROOT is empty " +
+                               help="If set, this script will make sure that STATIC_ROOT is empty " +
                                     "before running collectstatic by DELETING it (be careful!)")
-        gr_django.add_argument("--django-settings", dest="django_settings_module",
+        gr_django.add_argument("--django-settings", dest="django_settings_module", default=None,
                                help="'--settings' argument to pass to django-admin.py when it is called by " +
-                                    "this script")
+                                    "this script. If --django-generate-secret-key is set, SECRET_KEY will be set "
+                                    "in the environment.")
+        gr_django.add_argument("--django-gen-secret-key", dest="django_secret_key_file", default=None,
+                               env_var="DJANGO_GEN_SECRET_KEY",
+                               help="If set, GoPythonGo will write SECRET_KEY='(random)' to the given filename. The "
+                                    "resulting file can be read from envdir or systemd (EnvironmentFile). This is "
+                                    "useful for shipping environment configuration for projects adhering to 12factor "
+                                    "(and/or using the django12factor library).")
 
     def validate_args(self, args: configargparse.Namespace) -> None:
         if args.static_outfile or args.collect_static:
@@ -46,24 +54,33 @@ class DjangoAssembler(BaseAssembler):
                                    (highlight("--static-out"), highlight("--collect-static")))
 
     def assemble(self, args: configargparse.Namespace) -> None:
-        envpy = utils.create_script_path(args.build_path, "python")
-        print("Collecting static artifacts")
-        if os.path.exists(args.static_root):
-            print("    %s exists." % args.static_root)
-            if args.fresh_static:
-                print("removing stale static artifacts in %s" % args.static_root)
-                shutil.rmtree(args.static_root)
+        if args.django_secret_key_file:
+            secret = base64.b64encode(os.urandom(48))
+            if args.django_settings_module and 'SECRET_KEY' not in os.environ:
+                os.environ['SECRET_KEY'] = secret
 
-        django_admin = utils.create_script_path(args.build_path, 'django-admin.py')
-        run_dja = [envpy, django_admin, "collectstatic"]
-        if args.django_settings_module:
-            run_dja.append('--settings=%s' % args.django_settings_module)
-        run_dja.append("--noinput")
-        run_dja.append("--traceback")
-        utils.run_process(*run_dja)
+            with open(args.django_secret_key_file, "wt", encoding="utf-8") as sf:
+                sf.write("SECRET_KEY=\"%s\"\n" % secret)
 
-        if not os.path.exists(args.static_root):
-            raise ErrorMessage("%s should now exist, but it doesn't" % args.static_root)
+        if args.collect_static:
+            envpy = utils.create_script_path(args.build_path, "python")
+            print("Collecting static artifacts")
+            if os.path.exists(args.static_root):
+                print("    %s exists." % args.static_root)
+                if args.fresh_static:
+                    print("removing stale static artifacts in %s" % args.static_root)
+                    shutil.rmtree(args.static_root)
+
+            django_admin = utils.create_script_path(args.build_path, 'django-admin.py')
+            run_dja = [envpy, django_admin, "collectstatic"]
+            if args.django_settings_module:
+                run_dja.append('--settings=%s' % args.django_settings_module)
+            run_dja.append("--noinput")
+            run_dja.append("--traceback")
+            utils.run_process(*run_dja)
+
+            if not os.path.exists(args.static_root):
+                raise ErrorMessage("%s should now exist, but it doesn't" % args.static_root)
 
 
 assembler_class = DjangoAssembler  # type: Type[DjangoAssembler]
